Moved logic before React.
